requirements.txt
marshmallow==3.21.2
packaging==24.0
paho-mqtt==1.6.1
